+++ front-react/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Renamed `src/layouts/Dashboard/Dashboard.jsx` to `src/layouts/Admin.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Dashboard.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Icons.jsx`
-- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Maps.jsx`
+- Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Quiz.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Notifications.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Portfolio.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Youtube.jsx`